Resultado do Último Concurso

Utilize esta função para receber um conjunto de dados contendo todos o resultado do último concurso do produto escolhido.

Usage

resultado_ultimo_concurso(produto = "megasena")

Arguments

produto

Nome do jogo das Loterias que deseja atualizar a base. Caso não informado, por padrão será a Megasena megasena. Produtos disponíveis:

Megasena -> digite 'megasena' megasena

Lotofácil -> digite 'lotofacil' lotofacil

Quina -> digite 'quina' quina

Lotomania -> digite 'lotomania' lotomania

Timemania -> digite 'timemania' timemania

Super Sete -> digite 'supersete' supersete

Dia de Sorte -> digite 'diadesorte' diadesorte

Value

Uma tibble - a quantidade de variáveis depende do produto.

Details

Importante: o pacote atualiza as bases diariamente no repositório do GitHub, porém, sua base local permanece com os dados do momento em que o pacote foi instalado. Caso tenha intenção de usar as bases diretamente, sem chamar as funções, reinstale o pacote para obter as bases mais recentes.

Examples

# Visualizar concurso, números sorteados e se houve ganhador referente ao
# resultado do concurso mais recente da Megasena
resultado_ultimo_concurso(produto = 'megasena') |>
   dplyr::select(concurso, numeros_sorteados, houve_ganhador)
#> # A tibble: 1 × 3
#>   concurso numeros_sorteados houve_ganhador
#>      <dbl> <chr>                      <dbl>
#> 1     2885 10;25;28;36;37;56              0